requirements-linux.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0
python-dotenv==1.0.1
mysqlclient==2.2.4
google-auth-oauthlib==1.2.0
google-auth==2.28.1
Werkzeug==3.0.1
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.1
python-dateutil==2.8.2
Flask-Migrate==4.0.5
flask-caching==2.1.0
pymysql==1.1.0
gunicorn==21.2.0
greenlet==3.0.3
cryptography==42.0.5 